refactor: split gateway telemetry module
